+++ include/tensor_graph/tensor_graph.hpp
@@ -1,6 +1,0 @@
-#ifndef CAST_TENSOR_GRAPH_
-#define CAST_TENSOR_GRAPH_
-
-#include "tensor_operator.hpp"
-
-#endif